Properly added Genre playlist replacement.
Completely changed how the names are set so they use SetPanel instead of GetWordListInfo. This allows for genre playlists to be renamed without renaming genres entirely.
Also set PlaylistData.Name to string.Empty by default, so if it isn't set in the json, it will maintain the default name of the playlist.