Sorting implemented
I will do further testing to make sure everything's good.
I also need to update the example json, and I want to change how that's set up as well.
I want it to always recreate an example json next to the live json to make it easier to make updates to the json file and have users see the updates.